#317 change back the main executable to src/server/index.js
